// src/lib/api-zod/api.ts
export const UpdateUserBody = zod.object({
  name: zod.string().optional(),
  area: zod.string().nullish(),
  salesTarget: zod.number().nullish(),
  status: zod.enum(["active", "pending_interview", "rejected"]).optional(),
});